backend/portfolio-api/requirements.txt
boto3==1.18.58
botocore==1.21.58
jmespath==0.10.0
python-dateutil==2.8.1
s3transfer==0.5.0
six==1.15.0
urllib3==1.26.2

mock~=4.0.3
pip~=21.0
setuptools~=52.0.0

redis~=3.5.3
pytz~=2021.1
requests~=2.26.0

aws-lambda-powertools==1.21.1